refactor(12): Разделены клавиатуры
