+++ keyboards/inline_keyboards/other_kb.py
@@ -1,0 +1,16 @@
+from keyboards.keyboard_utils.inline_kb_utils import create_inline_kb
+
+main_menu_inline = create_inline_kb(
+    2,
+    pets_menu='Питомцы',
+    company='Компания',
+    pets='Кормления',
+    profile='Профиль',
+    supports='Поддержка',
+    about_bot='О боте',
+    user_offer='Пользовательское соглашение',
+)
+
+back_to_main_menu = create_inline_kb(
+    1, back_to_main_menu='⬅ Назад'
+)
